requirements.txt
# Core ML Libraries
scikit-learn==1.3.2
pandas==1.5.3
numpy==1.24.3

# Data Validation


# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Monitoring
prometheus-client==0.19.0

# Airflow


# Data Versioning
dvc[s3]==3.32.0
boto3==1.34.0

# Utilities
pyyaml==6.0.1
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.48

# Logging
python-json-logger==2.0.7